minor cleanup
- make 'mults_possible' extern in globalvars.h
- fix type inconsistencies
- drop size declaration from external array 'logfile'
- drop size declaration from external array 'buffer'
- drop size declaration from external array 'pr_hostaddress'
- drop trailing spaces

+++ src/main.c
@@ -286,7 +286,7 @@ int rig_comm_success = 0;
 int simulator = 0;
 int simulator_mode = 0;
 int simulator_seed = 8327;
-long system_secs;
+int system_secs;
 char tonecpy[5];
 char simulator_tone[5];
